+++ src/deus/index.ts
@@ -5,3 +5,4 @@ export * from "./debugOverlay";
 export * from "./fromTable";
 export * from "./templateTable";
 export * from "./scopedTransitions";
+export * from "./workflow";